Improve structure by separating entities into files
