--- Sources/InstantSearchCore/Searcher/AnswersSearcher.swift
@@ -1,10 +1,0 @@
-//
-//  AnswersSearcher.swift
-//  
-//
-//  Created by Vladislav Fitc on 30/11/2020.
-//
-
-import Foundation
-
-public final class AnswersSearcher: IndexSearcher<AlgoliaAnswersSearchService> {}
