+++ Sources/InstantSearchCore/Searcher/BaseSearcher.swift
@@ -120,65 +120,3 @@ public extension BaseSearcher {
   }
   
 }
-
-public protocol IndexNameProvider {
-  
-  var indexName: IndexName { get set }
-  
-}
-
-public protocol TextualQueryProvider {
-  
-  var textualQuery: String? { get set }
-  
-}
-
-
-public class IndexSearcher<Service: SearchService>: BaseSearcher<Service> where Service.Process == Operation, Service.Request: IndexNameProvider {
-  
-  public override var request: Request {
-    didSet {
-      if request.indexName != oldValue.indexName {
-        onIndexChanged.fire(request.indexName)
-      }
-    }
-  }
-
-  /// Triggered when an index of Searcher changed
-  /// - Parameter: equals to a new index value
-  public let onIndexChanged: Observer<IndexName>
-  
-  public override init(service: Service, initialRequest: Request) {
-    onIndexChanged = .init()
-    super.init(service: service, initialRequest: initialRequest)
-  }
-  
-}
-
-extension IndexSearcher: IndexNameSwitchable {
-  
-  public func switchIndexName(to indexName: IndexName) {
-    request.indexName = indexName
-  }
-  
-}
-
-
-public extension BaseSearcher where Service.Request: TextualQueryProvider {
-  
-  var query: String? {
-    get {
-      request.textualQuery
-    }
-
-    set {
-      let oldValue = request.textualQuery
-      request.textualQuery = newValue
-      if oldValue != newValue {
-        onQueryChanged.fire(newValue)
-      }
-    }
-  }
-
-  
-}
