+++ Sources/InstantSearchCore/Searcher/IndexNameProvider.swift
@@ -1,0 +1,14 @@
+//
+//  IndexNameProvider.swift
+//  
+//
+//  Created by Vladislav Fitc on 15/12/2020.
+//
+
+import Foundation
+
+public protocol IndexNameProvider {
+  
+  var indexName: IndexName { get set }
+  
+}
