+++ Sources/InstantSearchCore/Searcher/TextualQueryProvider.swift
@@ -1,0 +1,14 @@
+//
+//  TextualQueryProvider.swift
+//  
+//
+//  Created by Vladislav Fitc on 15/12/2020.
+//
+
+import Foundation
+
+public protocol TextualQueryProvider {
+  
+  var textualQuery: String? { get set }
+  
+}
